post/requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
kafka-python == 2.0.1
PyYAML >= 5.3.1